Changed user.getMeals design on JPA and Rest
