--- v1/Calories/WebContent/META-INF/context.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Context>
-   <Manager pathname=""/> <!-- disables storage of sessions across restarts -->
-   <Resource name="BeanManager"
-      auth="Container"
-      type="javax.enterprise.inject.spi.BeanManager"
-      factory="org.jboss.weld.resources.ManagerObjectFactory"/>
-</Context>
