+++ v1/Calories/WebContent/WEB-INF/web.xml
@@ -13,17 +13,6 @@
 		<welcome-file>default.jsp</welcome-file>
 	</welcome-file-list>
 
-	<listener>
-	   <listener-class>org.jboss.weld.environment.servlet.Listener</listener-class>
-	</listener>
-
-	<resource-env-ref>
-	   <resource-env-ref-name>BeanManager</resource-env-ref-name>
-	   <resource-env-ref-type>
-	      javax.enterprise.inject.spi.BeanManager
-	   </resource-env-ref-type>
-	</resource-env-ref>	
-	
 	<servlet>
 		<servlet-name>Jersey Web Application</servlet-name>
 		<servlet-class>com.sun.jersey.spi.container.servlet.ServletContainer</servlet-class>
