+++ v1/Calories/pom.xml
@@ -42,11 +42,6 @@
   		<version>1.19.1</version>
   	</dependency>
   	<dependency>
-  		<groupId>com.sun.jersey</groupId>
-  		<artifactId>jersey-core</artifactId>
-  		<version>1.19.1</version>
-  	</dependency>
-  	<dependency>
   		<groupId>org.json</groupId>
   		<artifactId>json</artifactId>
   		<version>20151123</version>
@@ -62,9 +57,9 @@
   		<version>1.3</version>
   	</dependency>
   	<dependency>
-  		<groupId>org.jboss.weld.servlet</groupId>
-  		<artifactId>weld-servlet</artifactId>
-  		<version>2.3.3.Final</version>
+  		<groupId>com.sun.jersey</groupId>
+  		<artifactId>jersey-core</artifactId>
+  		<version>1.19.1</version>
   	</dependency>
   </dependencies>
 </project>
