+++ v1/CaloriesManagementJPA/src/main/java/META-INF/persistence.xml
@@ -25,15 +25,13 @@ xsi:schemaLocation="http://java.sun.com/xml/ns/persistence file:/D:/Estudo/works
 			<property name="hibernate.show_sql" value="true" />
 			<property name="hibernate.format_sql" value="true" />
 			<property name="use_sql_comments" value="true" />
-			-->
+			 -->
 			<!--  End of Debug Configuration -->
 			
-			<!-- DEV SERVER -->
 			<property name="hibernate.connection.driver_class" value="org.postgresql.Driver" />
 			<property name="hibernate.connection.url" value="jdbc:postgresql://127.0.0.1:5432/toptal_calories" />
 			<property name="hibernate.connection.username" value="toptal_calories" />
 			<property name="hibernate.connection.password" value="password" />
-			<!--  -->
 						
 			<property name="hibernate.dialect" value="org.hibernate.dialect.PostgreSQLDialect"/>
 			<property name="hibernate.connection.provider_class" value="org.hibernate.connection.C3P0ConnectionProvider" />
