+++ v1/CaloriesManagementJPA/src/main/java/com/toptal/calories/resources/entity/Role.java
@@ -7,16 +7,24 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.NamedQuery;
 import javax.persistence.SequenceGenerator;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlType;
 
 
 /**
  * The persistent class for the role database table.
  * 
  */
+
+@XmlRootElement(name = "user")
+@XmlType(propOrder = {"id", "name"})
+@XmlAccessorType(XmlAccessType.FIELD)
+
 @Entity
 @NamedQuery(name="Role.findAll", query="SELECT r FROM Role r")
 public class Role extends BaseEntity {
-	//private static final long serialVersionUID = 1L;
 
 	@Id
 	@SequenceGenerator(name="role_id_seq", sequenceName="role_id_seq", allocationSize=1)
@@ -27,8 +35,8 @@ public class Role extends BaseEntity {
 	@Column(nullable=false)
 	private String name;
 
-	//bi-directional many-to-many association to User
 	/** not really needed
+	//bi-directional many-to-many association to User
 	@ManyToMany
 	@JoinTable(
 		name="user_role"
