+++ v1/CaloriesManagementJPA/src/test/java/com/toptal/calories/resources/TestDBBase.java
@@ -2,6 +2,7 @@ package com.toptal.calories.resources;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.security.InvalidParameterException;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
@@ -21,6 +22,39 @@ public class TestDBBase {
 		cal.add(unit, amount);
 		return  cal.getTime();
 	}
+	
+	public Date getDateDaysAgoAtTime(int daysAgo, int hour, int minute) {
+		Calendar cal = Calendar.getInstance();
+		cal.add(Calendar.DATE, -daysAgo);
+		
+		cal.set(Calendar.HOUR_OF_DAY, hour);
+		cal.set(Calendar.MINUTE, minute);
+		cal.set(Calendar.SECOND, 0);
+		cal.set(Calendar.MILLISECOND, 0);
+		
+		return cal.getTime();
+	}
+	
+	public Date getTime(int hours, int minutes) 
+	throws InvalidParameterException {
+		if (hours < 0 || hours > 23)
+			throw new InvalidParameterException("Invalid hour: " + hours);
+		if (minutes < 0 || minutes > 59)
+			throw new InvalidParameterException("Invalid minute: " + minutes);
+		
+		Calendar cal = Calendar.getInstance();
+		cal.set(Calendar.DATE, 1);
+		cal.set(Calendar.MONTH, 0);
+		cal.set(Calendar.YEAR, 1970);
+		cal.set(Calendar.HOUR_OF_DAY, hours);
+		cal.set(Calendar.MINUTE, minutes);
+		cal.set(Calendar.SECOND, 0);
+		cal.set(Calendar.MILLISECOND, 0);
+		
+		return cal.getTime();
+	}
+	
+	
 	
 	public String getCauseMessages(Throwable t) {
 		StringWriter sw = new StringWriter();
