+++ v1/CaloriesManagementJPA/src/test/java/com/toptal/calories/resources/repository/TestMeals.java
@@ -6,10 +6,7 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-import java.security.InvalidParameterException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
@@ -36,8 +33,6 @@ public class TestMeals extends TestDBBase {
 	private static User testUser;
 	private static String otherLogin = "2_" + login.substring(2);
 	private static User otherTestUser;
-	private static String thirdLogin = "3_" + login.substring(2);
-	private static User thirdTestUser;
 	
 	SimpleDateFormat dateOnly = new SimpleDateFormat("yyyy-MM-dd");
 	Date mealDate1;
@@ -48,8 +43,6 @@ public class TestMeals extends TestDBBase {
 
 	private static int invalidId = -1;
 
-	private List<Meal> dateTimeRangesTestMeals = new ArrayList<>();
-	
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
 		logger.debug("Creating test users to be used by all tests...");
@@ -60,9 +53,6 @@ public class TestMeals extends TestDBBase {
 		otherTestUser = TestUsers.getUser(otherLogin);
 		otherTestUser = TestUsers.createUser(otherTestUser);
 
-		thirdTestUser = TestUsers.getUser(thirdLogin);
-		thirdTestUser = TestUsers.createUser(thirdTestUser);
-		
 		logger.debug("Finished creating test users to be used by all tests");
 	}
 
@@ -86,7 +76,6 @@ public class TestMeals extends TestDBBase {
 		
 		TestUsers.removeUser(testUser);
 		TestUsers.removeUser(otherTestUser);
-		TestUsers.removeUser(thirdTestUser);
 
 		logger.debug("Finished deleting test users after all tests done");
 	}
@@ -98,7 +87,7 @@ public class TestMeals extends TestDBBase {
 	public static Meal getMeal(User user, Date date, Date time) {
 		
 		Meal meal = new Meal();
-		meal.setUserId(user.getId());
+		meal.setUser(user);
 		meal.setMealDate(date);
 		meal.setMealTime(time);
 		meal.setDescription("Test description");
@@ -246,18 +235,18 @@ public class TestMeals extends TestDBBase {
 	 * Tests null Meal User Id sent for creation 
 	 */
 	@Test
-	public void testCreateMealNullUserId() {
+	public void testCreateMealNullUser() {
 		logger.debug("Running " + getCurrentMethodName());
 		try {
 			Meal meal = getMeal(testUser, mealDate1, mealTime1);
-			meal.setUserId(null);
+			meal.setUser(null);
 			repository.createOrUpdate(meal);
-			fail("Null User ID should have caused a RepositoryExcpetion!");
+			fail("Null User should have caused a RepositoryExcpetion!");
 		} catch (RepositoryException re) {
 			// All good
 		} catch (Exception e) {
 			printException(e);
-			fail("Null User ID should have caused a RepositoryExcpetion!");
+			fail("Null User should have caused a RepositoryExcpetion!");
 		}
 	}
 
@@ -421,584 +410,4 @@ public class TestMeals extends TestDBBase {
 			fail("Invalid Meal ID should NOT have caused a RepositoryExcpetion!");
 		}
 	}
-	
-	/******************************************************************************/
-	/**  		            Extended queries tests                               **/											
-	/******************************************************************************/
-	
-	/******************************************************************************/
-	/** Find user meals tests                                                    **/
-	/******************************************************************************/
-	
-	@Test
-	public void testFindUserMealsNullUserId() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMeals(null);
-			fail("Null User ID should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null User ID should have caused a RepositoryExcpetion!");
-		}
-	}
-	
-	@Test
-	public void testFindUserMealsInvalidUser() {
-		logger.debug("Running " + getCurrentMethodName());
-		Meal testMeal = null;
-		Meal secondMeal = null;
-		Meal thirdMeal = null;
-		Meal fourthMeal = null;
-		try {
-			// create test Meal
-			testMeal = getMeal(testUser, mealDate1, mealTime1);
-			testMeal = repository.createOrUpdate(testMeal);
-			// create a second Meal
-			secondMeal = getMeal(testUser, mealDate2, mealTime2);
-			secondMeal = repository.createOrUpdate(secondMeal);
-			// create a third Meal
-			thirdMeal = getMeal(otherTestUser, mealDate1, mealTime1);
-			thirdMeal = repository.createOrUpdate(thirdMeal);
-			// create a fourth Meal
-			fourthMeal = getMeal(otherTestUser, mealDate2, mealTime2);
-			fourthMeal = repository.createOrUpdate(fourthMeal);
-			
-			List<Meal> meals = repository.findUserMeals(invalidId);
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals should NOT have caused an Exception!");
-		} finally {
-			try {
-				if (testMeal != null) {
-					repository.remove(testMeal.getId());
-				}
-				if (secondMeal != null) {
-					repository.remove(secondMeal.getId());
-				}
-				if (thirdMeal != null) {
-					repository.remove(thirdMeal.getId());
-				}
-				if (fourthMeal != null) {
-					repository.remove(fourthMeal.getId());
-				}
-			} catch (Exception e) {
-				logger.error("In Finally Block: Failed to remove entities", e);
-			}
-		}
-	}
-	
-	@Test
-	public void testFindUserMealsValidUserNoMeals() {
-		logger.debug("Running " + getCurrentMethodName());
-		Meal testMeal = null;
-		Meal secondMeal = null;
-		Meal thirdMeal = null;
-		Meal fourthMeal = null;
-		try {
-			// create test Meal
-			testMeal = getMeal(testUser, mealDate1, mealTime1);
-			testMeal = repository.createOrUpdate(testMeal);
-			// create a second Meal
-			secondMeal = getMeal(testUser, mealDate2, mealTime2);
-			secondMeal = repository.createOrUpdate(secondMeal);
-			// create a third Meal
-			thirdMeal = getMeal(otherTestUser, mealDate1, mealTime1);
-			thirdMeal = repository.createOrUpdate(thirdMeal);
-			// create a fourth Meal
-			fourthMeal = getMeal(otherTestUser, mealDate2, mealTime2);
-			fourthMeal = repository.createOrUpdate(fourthMeal);
-			
-			List<Meal> meals = repository.findUserMeals(invalidId);
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals should NOT have caused an Exception!");
-		} finally {
-			try {
-				if (testMeal != null) {
-					repository.remove(testMeal.getId());
-				}
-				if (secondMeal != null) {
-					repository.remove(secondMeal.getId());
-				}
-				if (thirdMeal != null) {
-					repository.remove(thirdMeal.getId());
-				}
-				if (fourthMeal != null) {
-					repository.remove(fourthMeal.getId());
-				}
-			} catch (Exception e) {
-				logger.error("In Finally Block: Failed to remove entities", e);
-			}
-		}
-	}
-	
-	
-	@Test
-	public void testFindUserMealsValidUserOneMeal() {
-		logger.debug("Running " + getCurrentMethodName());
-		Meal testMeal = null;
-		Meal secondMeal = null;
-		Meal thirdMeal = null;
-		Meal fourthMeal = null;
-		try {
-			// create test Meal
-			testMeal = getMeal(testUser, mealDate1, mealTime1);
-			testMeal = repository.createOrUpdate(testMeal);
-			// create a second Meal
-			secondMeal = getMeal(testUser, mealDate2, mealTime2);
-			secondMeal = repository.createOrUpdate(secondMeal);
-			// create a third Meal
-			thirdMeal = getMeal(otherTestUser, mealDate1, mealTime1);
-			thirdMeal = repository.createOrUpdate(thirdMeal);
-			
-			List<Meal> meals = repository.findUserMeals(otherTestUser.getId());
-			assertNotNull("One meal should have been found", meals);
-			assertEquals("One meal should have been found", 1, meals.size());
-			assertTrue("Created meals were not found in the returned list", meals.contains(thirdMeal));
-		} catch (RepositoryException re) {
-			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals should NOT have caused an Exception!");
-		} finally {
-			try {
-				if (testMeal != null) {
-					repository.remove(testMeal.getId());
-				}
-				if (secondMeal != null) {
-					repository.remove(secondMeal.getId());
-				}
-				if (thirdMeal != null) {
-					repository.remove(thirdMeal.getId());
-				}
-				if (fourthMeal != null) {
-					repository.remove(fourthMeal.getId());
-				}
-			} catch (Exception e) {
-				logger.error("In Finally Block: Failed to remove entities", e);
-			}
-		}
-	}
-	
-	@Test
-	public void testFindUserMealsValidUserTwoMeals() {
-		logger.debug("Running " + getCurrentMethodName());
-		Meal testMeal = null;
-		Meal secondMeal = null;
-		Meal thirdMeal = null;
-		Meal fourthMeal = null;
-		try {
-			// create test Meal
-			testMeal = getMeal(testUser, mealDate1, mealTime1);
-			testMeal = repository.createOrUpdate(testMeal);
-			// create a second Meal
-			secondMeal = getMeal(testUser, mealDate2, mealTime2);
-			secondMeal = repository.createOrUpdate(secondMeal);
-			// create a third Meal
-			thirdMeal = getMeal(otherTestUser, mealDate1, mealTime1);
-			thirdMeal = repository.createOrUpdate(thirdMeal);
-			// create a fourth Meal
-			fourthMeal = getMeal(otherTestUser, mealDate2, mealTime2);
-			fourthMeal = repository.createOrUpdate(fourthMeal);
-			
-			List<Meal> meals = repository.findUserMeals(testUser.getId());
-			assertNotNull("Two meals should have been found", meals);
-			assertEquals("Two meals should have been found", 2, meals.size());
-			assertTrue("Created meals were not found in the returned list", meals.contains(testMeal) && meals.contains(secondMeal));
-		} catch (RepositoryException re) {
-			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals should NOT have caused an Exception!");
-		} finally {
-			try {
-				if (testMeal != null) {
-					repository.remove(testMeal.getId());
-				}
-				if (secondMeal != null) {
-					repository.remove(secondMeal.getId());
-				}
-				if (thirdMeal != null) {
-					repository.remove(thirdMeal.getId());
-				}
-				if (fourthMeal != null) {
-					repository.remove(fourthMeal.getId());
-				}
-			} catch (Exception e) {
-				logger.error("In Finally Block: Failed to remove entities", e);
-			}
-		}
-	}
-	
-	/******************************************************************************/
-	/** Find user meals in date and time range tests                             **/
-	/******************************************************************************/
-
-	/******************************************************************************/
-	/** Test specific helper methods 	                                         **/
-	/******************************************************************************/
-	
-	private void prepareForDateTimeRangeTests() 
-	throws RepositoryException {
-		// create 3 meals a day for 2 users for the last 60 days
-		int numDays = 60;
-		int[] mealTimeHours = {9, 12, 18};
-		int[] mealTimeMinutes = {0, 30, 45};
-		User[] users = {testUser, otherTestUser};
-		
-		Date mealTime;
-		Meal meal;
-		
-		try {
-			for (User user: users) {
-				for (int i = numDays; i >= 0; i--) {
-					for (int j = 0; j < mealTimeHours.length; j++) {
-						mealTime = getDateDaysAgoAtTime(i, mealTimeHours[j], mealTimeMinutes[j]);
-						
-						meal = getMeal(user, mealTime, mealTime);
-						meal = createMeal(meal);
-						dateTimeRangesTestMeals.add(meal);
-					}
-				}
-			}
-		} catch (RepositoryException re) {
-			logger.error("Error creating test meals for date range test", re);
-
-			try {
-				for (Meal toRemove: dateTimeRangesTestMeals) {
-					removeMeal(toRemove);
-				}
-			} catch (RepositoryException re2) {
-				logger.error("Error removing test meals after creation failure", re2);
-			}
-			
-			throw re;
-		}
-	}
-	
-	private void afterDateTimeRangeTest() {
-		for (Meal toRemove : dateTimeRangesTestMeals) {
-			try {
-				removeMeal(toRemove);
-			} catch (RepositoryException re) {
-				logger.error("Error removing test meals", re);
-			}
-		}
-		dateTimeRangesTestMeals.clear();
-	}
-	
-	private Date getDateDaysAgoAtTime(int daysAgo, int hour, int minute) {
-		Calendar cal = Calendar.getInstance();
-		cal.add(Calendar.DATE, -daysAgo);
-		
-		cal.set(Calendar.HOUR_OF_DAY, hour);
-		cal.set(Calendar.MINUTE, minute);
-		cal.set(Calendar.SECOND, 0);
-		cal.set(Calendar.MILLISECOND, 0);
-		
-		return cal.getTime();
-	}
-	
-	private Date getTime(int hours, int minutes) 
-	throws InvalidParameterException {
-		if (hours < 0 || hours > 23)
-			throw new InvalidParameterException("Invalid hour: " + hours);
-		if (minutes < 0 || minutes > 59)
-			throw new InvalidParameterException("Invalid minute: " + minutes);
-		
-		Calendar cal = Calendar.getInstance();
-		cal.set(Calendar.DATE, 1);
-		cal.set(Calendar.MONTH, 0);
-		cal.set(Calendar.YEAR, 1970);
-		cal.set(Calendar.HOUR_OF_DAY, hours);
-		cal.set(Calendar.MINUTE, minutes);
-		cal.set(Calendar.SECOND, 0);
-		cal.set(Calendar.MILLISECOND, 0);
-		
-		return cal.getTime();
-	}
-	
-	/******************************************************************************/
-	/** Actual tests 	                                                         **/
-	/******************************************************************************/
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNullUser() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(null, new Date(0), new Date(), getTime(0,0), getTime(23,59));
-			fail("Null User ID should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null User ID should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNullFromDate() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), null, new Date(), getTime(0,0), getTime(23,59));
-			fail("Null start date should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null end date should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNullToDate() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), new Date(0), null, getTime(0,0), getTime(23,59));
-			fail("Null end date should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null end date should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNullFromTime() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), null, getTime(23,59));
-			fail("Null start time should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null start time  should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNullToTime() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), getTime(0,0), null);
-			fail("Null end time should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Null end time should have caused a RepositoryExcpetion!");
-		}
-	}
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesStartDateAfterEndDate() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// start date = today, end date = yesterday
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(0, 0, 0), getDateDaysAgoAtTime(1, 0, 0), getTime(0,0), getTime(23,59));
-			fail("Invalid start and end dates should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Invalid start and end dates should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesStartTimeAfterEndTime() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			repository.findUserMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), getTime(15,10), getTime(15,9));
-			fail("Invalid start and end times should have caused a RepositoryExcpetion!");
-		} catch (RepositoryException re) {
-			// All good
-		} catch (Exception e) {
-			printException(e);
-			fail("Invalid start and end times should have caused a RepositoryExcpetion!");
-		}
-	}
-
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesInvalidUser() {
-		logger.debug("Running " + getCurrentMethodName());
-		Meal testMeal = null;
-		Meal secondMeal = null;
-		Meal thirdMeal = null;
-		Meal fourthMeal = null;
-		try {
-			// create test Meal
-			testMeal = getMeal(testUser, mealDate1, mealTime1);
-			testMeal = repository.createOrUpdate(testMeal);
-			// create a second Meal
-			secondMeal = getMeal(testUser, mealDate2, mealTime2);
-			secondMeal = repository.createOrUpdate(secondMeal);
-			// create a third Meal
-			thirdMeal = getMeal(otherTestUser, mealDate1, mealTime1);
-			thirdMeal = repository.createOrUpdate(thirdMeal);
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(invalidId, new Date(0), new Date(), getTime(0,0), getTime(23,59));
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("testDateFuncsFourMeals should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("testDateFuncsFourMeals should NOT have caused an Exception!");
-		} finally {
-			try {
-				if (testMeal != null) {
-					repository.remove(testMeal.getId());
-				}
-				if (secondMeal != null) {
-					repository.remove(secondMeal.getId());
-				}
-				if (thirdMeal != null) {
-					repository.remove(thirdMeal.getId());
-				}
-				if (fourthMeal != null) {
-					repository.remove(fourthMeal.getId());
-				}
-			} catch (Exception e) {
-				logger.error("In Finally Block: Failed to remove entities", e);
-			}
-		}
-	}
-
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNoMealsForUser() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(thirdTestUser.getId(), new Date(0), new Date(), getTime(0,0), getTime(23,59));
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNoMealsForDateRange() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(62, 0, 0), getDateDaysAgoAtTime(61, 0, 0), getTime(0,0), getTime(23,59));
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesNoMealsForTimeRange() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(0,0), getTime(8,30));
-			assertNotNull("No meals should have been found", meals);
-			assertEquals("No meals should have been found", 0, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesStartEndDatesSame() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(60, 0, 0), getTime(0,0), getTime(23,59));
-			assertNotNull("Meals should have been found", meals);
-			assertEquals("Unexpected number of meals found", 3, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesStartEndTimesSame() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(9,0), getTime(9,0));
-			assertNotNull("Meals should have been found", meals);
-			assertEquals("Unexpected number of meals found", 61, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-	
-	@Test
-	public void testfindUserMealsInDateAndTimeRangesAllMeals() {
-		logger.debug("Running " + getCurrentMethodName());
-		try {
-			// create test meals
-			prepareForDateTimeRangeTests();
-			
-			List<Meal> meals = repository.findUserMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(0,0), getTime(23,0));
-			assertNotNull("Meals should have been found", meals);
-			assertEquals("Unexpected number of meals found", 183, meals.size());
-		} catch (RepositoryException re) {
-			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
-		} catch (Exception e) {
-			printException(e);
-			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
-		} finally {
-			afterDateTimeRangeTest();
-		}
-	}
-	
-
-	
 }
