+++ v1/CaloriesManagementJPA/src/test/java/com/toptal/calories/resources/repository/TestUsers.java
@@ -13,13 +13,16 @@ import java.util.Date;
 import java.util.List;
 
 import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.toptal.calories.resources.TestDBBase;
 import com.toptal.calories.resources.entity.Gender;
+import com.toptal.calories.resources.entity.Meal;
 import com.toptal.calories.resources.entity.Role;
 import com.toptal.calories.resources.entity.User;
 
@@ -33,16 +36,64 @@ public class TestUsers extends TestDBBase {
 	private static String login = sdf.format(new Date());
 	private static String otherLogin = "2_" + login.substring(2);
 
+	private static User testUser;
+	private static User otherTestUser;
+	private static String thirdLogin = "3_" + login.substring(2);
+	private static User thirdTestUser;
+	
 	private static int invalidId = -1;
 	
+	SimpleDateFormat dateOnly = new SimpleDateFormat("yyyy-MM-dd");
+	Date mealDate1;
+	Date mealTime1;
+
+	Date mealDate2;
+	Date mealTime2;
+
+	private List<Meal> dateTimeRangesTestMeals = new ArrayList<>();
+	
+	@BeforeClass
+	public static void setUpBeforeClass() throws Exception {
+		logger.debug("Creating test users to be used by all tests...");
+		
+		testUser = TestUsers.getUser(login);
+		testUser = TestUsers.createUser(testUser);
+		
+		otherTestUser = TestUsers.getUser(otherLogin);
+		otherTestUser = TestUsers.createUser(otherTestUser);
+
+		thirdTestUser = TestUsers.getUser(thirdLogin);
+		thirdTestUser = TestUsers.createUser(thirdTestUser);
+		
+		logger.debug("Finished creating test users to be used by all tests");
+	}
+
 	@Before
 	public void setUp() throws Exception {
+		SimpleDateFormat dateOnly = new SimpleDateFormat("yyyy-MM-dd");
+		mealDate1 = dateOnly.parse(dateOnly.format(new Date()));
+		mealTime1 = getTime(0, 0);
+
+		mealDate2 = dateOnly.parse(dateOnly.format(new Date()));
+		mealTime2 = getTime(23, 59);
 	}
 
 	@After
 	public void tearDown() throws Exception {
 	}
 
+	@AfterClass
+	public static void afterClass() throws Exception {
+		logger.debug("Deleting test users after all tests done...");
+		
+		TestUsers.removeUser(testUser);
+		TestUsers.removeUser(otherTestUser);
+		TestUsers.removeUser(thirdTestUser);
+
+		logger.debug("Finished deleting test users after all tests done");
+	}
+	
+	
 	/******************************************************************************/
 	/**  		            Helper Methods                                       **/											
 	/******************************************************************************/
@@ -792,4 +843,566 @@ public class TestUsers extends TestDBBase {
 			}
 		}
 	}	
+	
+	/******************************************************************************/
+	/** Find user meals tests                                                    **/
+	/******************************************************************************/
+	
+	@Test
+	public void testFindUserMealsNullUserId() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.find(null);
+			fail("Null User ID should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null User ID should have caused a RepositoryExcpetion!");
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsInvalidUser() {
+		logger.debug("Running " + getCurrentMethodName());
+		Meal testMeal = null;
+		Meal secondMeal = null;
+		Meal thirdMeal = null;
+		Meal fourthMeal = null;
+		try {
+			// create test Meal
+			testMeal = TestMeals.getMeal(testUser, mealDate1, mealTime1);
+			testMeal = TestMeals.createMeal(testMeal);
+			// create a second Meal
+			secondMeal = TestMeals.getMeal(testUser, mealDate2, mealTime2);
+			secondMeal = TestMeals.createMeal(secondMeal);
+			// create a third Meal
+			thirdMeal = TestMeals.getMeal(otherTestUser, mealDate1, mealTime1);
+			thirdMeal = TestMeals.createMeal(thirdMeal);
+			// create a fourth Meal
+			fourthMeal = TestMeals.getMeal(otherTestUser, mealDate2, mealTime2);
+			fourthMeal = TestMeals.createMeal(fourthMeal);
+
+			// find the user owning the meals
+			User user = repository.find(invalidId);
+			assertNull("User should be null", user);
+		} catch (RepositoryException re) {
+			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals should NOT have caused an Exception!");
+		} finally {
+			try {
+				if (testMeal != null) {
+					TestMeals.removeMeal(testMeal);
+				}
+				if (secondMeal != null) {
+					TestMeals.removeMeal(secondMeal);
+				}
+				if (thirdMeal != null) {
+					TestMeals.removeMeal(thirdMeal);
+				}
+				if (fourthMeal != null) {
+					TestMeals.removeMeal(fourthMeal);
+				}
+			} catch (Exception e) {
+				logger.error("In Finally Block: Failed to remove entities", e);
+			}
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsValidUserNoMeals() {
+		logger.debug("Running " + getCurrentMethodName());
+		Meal testMeal = null;
+		Meal secondMeal = null;
+		Meal thirdMeal = null;
+		Meal fourthMeal = null;
+		try {
+			// create test Meal
+			testMeal = TestMeals.getMeal(testUser, mealDate1, mealTime1);
+			testMeal = TestMeals.createMeal(testMeal);
+			// create a second Meal
+			secondMeal = TestMeals.getMeal(testUser, mealDate2, mealTime2);
+			secondMeal = TestMeals.createMeal(secondMeal);
+			// create a third Meal
+			thirdMeal = TestMeals.getMeal(otherTestUser, mealDate1, mealTime1);
+			thirdMeal = TestMeals.createMeal(thirdMeal);
+			// create a fourth Meal
+			fourthMeal = TestMeals.getMeal(otherTestUser, mealDate2, mealTime2);
+			fourthMeal = TestMeals.createMeal(fourthMeal);
+			
+			User user = repository.find(thirdTestUser.getId());
+			assertNotNull("Valid user should have been found", user);
+			
+			List<Meal> meals = user.getMeals();
+			assertNotNull("No meals should have been found", meals);
+			assertEquals("No meals should have been found", 0, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals should NOT have caused an Exception!");
+		} finally {
+			try {
+				if (testMeal != null) {
+					TestMeals.removeMeal(testMeal);
+				}
+				if (secondMeal != null) {
+					TestMeals.removeMeal(secondMeal);
+				}
+				if (thirdMeal != null) {
+					TestMeals.removeMeal(thirdMeal);
+				}
+				if (fourthMeal != null) {
+					TestMeals.removeMeal(fourthMeal);
+				}
+			} catch (Exception e) {
+				logger.error("In Finally Block: Failed to remove entities", e);
+			}
+		}
+	}
+	
+	
+	@Test
+	public void testFindUserMealsValidUserOneMeal() {
+		logger.debug("Running " + getCurrentMethodName());
+		Meal testMeal = null;
+		Meal secondMeal = null;
+		Meal thirdMeal = null;
+		Meal fourthMeal = null;
+		try {
+			// create test Meal
+			testMeal = TestMeals.getMeal(testUser, mealDate1, mealTime1);
+			testMeal = TestMeals.createMeal(testMeal);
+			// create a second Meal
+			secondMeal = TestMeals.getMeal(testUser, mealDate2, mealTime2);
+			secondMeal = TestMeals.createMeal(secondMeal);
+			// create a third Meal
+			thirdMeal = TestMeals.getMeal(otherTestUser, mealDate1, mealTime1);
+			thirdMeal = TestMeals.createMeal(thirdMeal);
+			
+			User user = repository.find(otherTestUser.getId());
+			assertNotNull("Valid user should have been found", user);
+			
+			List<Meal> meals = user.getMeals();
+			assertNotNull("One meal should have been found", meals);
+			assertEquals("One meal should have been found", 1, meals.size());
+			assertTrue("Created meals were not found in the returned list", meals.contains(thirdMeal));
+		} catch (RepositoryException re) {
+			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals should NOT have caused an Exception!");
+		} finally {
+			try {
+				if (testMeal != null) {
+					TestMeals.removeMeal(testMeal);
+				}
+				if (secondMeal != null) {
+					TestMeals.removeMeal(secondMeal);
+				}
+				if (thirdMeal != null) {
+					TestMeals.removeMeal(thirdMeal);
+				}
+				if (fourthMeal != null) {
+					TestMeals.removeMeal(fourthMeal);
+				}
+			} catch (Exception e) {
+				logger.error("In Finally Block: Failed to remove entities", e);
+			}
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsValidUserTwoMeals() {
+		logger.debug("Running " + getCurrentMethodName());
+		Meal testMeal = null;
+		Meal secondMeal = null;
+		Meal thirdMeal = null;
+		Meal fourthMeal = null;
+		try {
+			// create test Meal
+			testMeal = TestMeals.getMeal(testUser, mealDate1, mealTime1);
+			testMeal = TestMeals.createMeal(testMeal);
+			// create a second Meal
+			secondMeal = TestMeals.getMeal(testUser, mealDate2, mealTime2);
+			secondMeal = TestMeals.createMeal(secondMeal);
+			// create a third Meal
+			thirdMeal = TestMeals.getMeal(otherTestUser, mealDate1, mealTime1);
+			thirdMeal = TestMeals.createMeal(thirdMeal);
+			// create a fourth Meal
+			fourthMeal = TestMeals.getMeal(otherTestUser, mealDate2, mealTime2);
+			fourthMeal = TestMeals.createMeal(fourthMeal);
+			
+			User user = repository.find(testUser.getId());
+			assertNotNull("Valid user should have been found", user);
+			
+			List<Meal> meals = user.getMeals();
+			assertNotNull("Two meals should have been found", meals);
+			assertEquals("Two meals should have been found", 2, meals.size());
+			assertTrue("Created meals were not found in the returned list", meals.contains(testMeal) && meals.contains(secondMeal));
+		} catch (RepositoryException re) {
+			fail("Find User meals should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals should NOT have caused an Exception!");
+		} finally {
+			try {
+				if (testMeal != null) {
+					TestMeals.removeMeal(testMeal);
+				}
+				if (secondMeal != null) {
+					TestMeals.removeMeal(secondMeal);
+				}
+				if (thirdMeal != null) {
+					TestMeals.removeMeal(thirdMeal);
+				}
+				if (fourthMeal != null) {
+					TestMeals.removeMeal(fourthMeal);
+				}
+			} catch (Exception e) {
+				logger.error("In Finally Block: Failed to remove entities", e);
+			}
+		}
+	}
+	
+	/******************************************************************************/
+	/** Find user meals in date and time range tests                             **/
+	/******************************************************************************/
+
+	/******************************************************************************/
+	/** Test specific helper methods 	                                         **/
+	/******************************************************************************/
+	
+	private void prepareForDateTimeRangeTests() 
+	throws RepositoryException {
+		// create 3 meals a day for 2 users for the last 60 days
+		int numDays = 60;
+		int[] mealTimeHours = {9, 12, 18};
+		int[] mealTimeMinutes = {0, 30, 45};
+		User[] users = {testUser, otherTestUser};
+		
+		Date mealTime;
+		Meal meal;
+		
+		try {
+			for (User user: users) {
+				for (int i = numDays; i >= 0; i--) {
+					for (int j = 0; j < mealTimeHours.length; j++) {
+						mealTime = getDateDaysAgoAtTime(i, mealTimeHours[j], mealTimeMinutes[j]);
+						
+						meal = TestMeals.getMeal(user, mealTime, mealTime);
+						meal = TestMeals.createMeal(meal);
+						dateTimeRangesTestMeals.add(meal);
+					}
+				}
+			}
+		} catch (RepositoryException re) {
+			logger.error("Error creating test meals for date range test", re);
+
+			try {
+				for (Meal toRemove: dateTimeRangesTestMeals) {
+					TestMeals.removeMeal(toRemove);
+				}
+			} catch (RepositoryException re2) {
+				logger.error("Error removing test meals after creation failure", re2);
+			}
+			
+			throw re;
+		}
+	}
+	
+	private void afterDateTimeRangeTest() {
+		for (Meal toRemove : dateTimeRangesTestMeals) {
+			try {
+				TestMeals.removeMeal(toRemove);
+			} catch (RepositoryException re) {
+				logger.error("Error removing test meals", re);
+			}
+		}
+		dateTimeRangesTestMeals.clear();
+	}
+	
+	/******************************************************************************/
+	/** Actual tests 	                                                         **/
+	/******************************************************************************/
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNullUser() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(null, new Date(0), new Date(), getTime(0,0), getTime(23,59));
+			fail("Null User ID should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null User ID should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNullFromDate() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), null, new Date(), getTime(0,0), getTime(23,59));
+			fail("Null start date should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null end date should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNullToDate() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), new Date(0), null, getTime(0,0), getTime(23,59));
+			fail("Null end date should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null end date should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNullFromTime() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), null, getTime(23,59));
+			fail("Null start time should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null start time  should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNullToTime() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), getTime(0,0), null);
+			fail("Null end time should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Null end time should have caused a RepositoryExcpetion!");
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesStartDateAfterEndDate() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// start date = today, end date = yesterday
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(0, 0, 0), getDateDaysAgoAtTime(1, 0, 0), getTime(0,0), getTime(23,59));
+			fail("Invalid start and end dates should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Invalid start and end dates should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesStartTimeAfterEndTime() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			repository.findMealsInDateAndTimeRanges(testUser.getId(), new Date(0), new Date(), getTime(15,10), getTime(15,9));
+			fail("Invalid start and end times should have caused a RepositoryExcpetion!");
+		} catch (RepositoryException re) {
+			// All good
+		} catch (Exception e) {
+			printException(e);
+			fail("Invalid start and end times should have caused a RepositoryExcpetion!");
+		}
+	}
+
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesInvalidUser() {
+		logger.debug("Running " + getCurrentMethodName());
+		Meal testMeal = null;
+		Meal secondMeal = null;
+		Meal thirdMeal = null;
+		Meal fourthMeal = null;
+		try {
+			// create test Meal
+			testMeal = TestMeals.getMeal(testUser, mealDate1, mealTime1);
+			testMeal = TestMeals.createMeal(testMeal);
+			// create a second Meal
+			secondMeal = TestMeals.getMeal(testUser, mealDate2, mealTime2);
+			secondMeal = TestMeals.createMeal(secondMeal);
+			// create a third Meal
+			thirdMeal = TestMeals.getMeal(otherTestUser, mealDate1, mealTime1);
+			thirdMeal = TestMeals.createMeal(thirdMeal);
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(invalidId, new Date(0), new Date(), getTime(0,0), getTime(23,59));
+			assertNotNull("No meals should have been found", meals);
+			assertEquals("No meals should have been found", 0, meals.size());
+		} catch (RepositoryException re) {
+			fail("testDateFuncsFourMeals should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("testDateFuncsFourMeals should NOT have caused an Exception!");
+		} finally {
+			try {
+				if (testMeal != null) {
+					TestMeals.removeMeal(testMeal);
+				}
+				if (secondMeal != null) {
+					TestMeals.removeMeal(secondMeal);
+				}
+				if (thirdMeal != null) {
+					TestMeals.removeMeal(thirdMeal);
+				}
+				if (fourthMeal != null) {
+					TestMeals.removeMeal(fourthMeal);
+				}
+			} catch (Exception e) {
+				logger.error("In Finally Block: Failed to remove entities", e);
+			}
+		}
+	}
+
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNoMealsForUser() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(thirdTestUser.getId(), new Date(0), new Date(), getTime(0,0), getTime(23,59));
+			assertNotNull("No meals should have been found", meals);
+			assertEquals("No meals should have been found", 0, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNoMealsForDateRange() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(62, 0, 0), getDateDaysAgoAtTime(61, 0, 0), getTime(0,0), getTime(23,59));
+			assertNotNull("No meals should have been found", meals);
+			assertEquals("No meals should have been found", 0, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesNoMealsForTimeRange() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(0,0), getTime(8,30));
+			assertNotNull("No meals should have been found", meals);
+			assertEquals("No meals should have been found", 0, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesStartEndDatesSame() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(60, 0, 0), getTime(0,0), getTime(23,59));
+			assertNotNull("Meals should have been found", meals);
+			assertEquals("Unexpected number of meals found", 3, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesStartEndTimesSame() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(9,0), getTime(9,0));
+			assertNotNull("Meals should have been found", meals);
+			assertEquals("Unexpected number of meals found", 61, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+	
+	@Test
+	public void testFindUserMealsInDateAndTimeRangesAllMeals() {
+		logger.debug("Running " + getCurrentMethodName());
+		try {
+			// create test meals
+			prepareForDateTimeRangeTests();
+			
+			List<Meal> meals = repository.findMealsInDateAndTimeRanges(testUser.getId(), getDateDaysAgoAtTime(60, 0, 0), getDateDaysAgoAtTime(0, 0, 0), getTime(0,0), getTime(23,0));
+			assertNotNull("Meals should have been found", meals);
+			assertEquals("Unexpected number of meals found", 183, meals.size());
+		} catch (RepositoryException re) {
+			fail("Find User meals In Date And Time Ranges should NOT have caused a RepositoryExcpetion!");
+		} catch (Exception e) {
+			printException(e);
+			fail("Find User meals In Date And Time Ranges should NOT have caused an Exception!");
+		} finally {
+			afterDateTimeRangeTest();
+		}
+	}
+	
+
+	
+	
+	
+//	public static void main(String[] args) throws Exception {
+//		User user = repository.find(1157);
+//		System.out.println(user);
+//
+//		System.out.println("User Meals: " + (user != null ? user.getMeals() : "null"));
+//	}
 }
